Wrap whole backlink with a class
I want to be able to style/replace the backlink as needed, on my summary pages I just want to restyle it with a icon like the edit button that takes you to the page it was created on but without all the space taken by the text. This is the most straight forward way I could think of as well as allowing others to style the content.